Refactoring of taar cache to provide independent context for Ensemble spark job
