+++ bin/test_env.sh
@@ -4,3 +4,19 @@ export TAAR_MAX_RESULTS=10
 export BIGTABLE_PROJECT_ID=moz-fx-data-taar-pr-prod-e0f7
 export BIGTABLE_INSTANCE_ID=taar-prod-202006
 export BIGTABLE_TABLE_ID=taar_profile
+export TAAR_ITEM_MATRIX_BUCKET=moz-fx-data-taar-pr-prod-e0f7-prod-models
+export TAAR_ITEM_MATRIX_KEY=addon_recommender/item_matrix.json.bz2
+export TAAR_ADDON_MAPPING_BUCKET=moz-fx-data-taar-pr-prod-e0f7-prod-models
+export TAAR_ADDON_MAPPING_KEY=addon_recommender/addon_mapping.json.bz2
+export TAAR_ENSEMBLE_BUCKET=taar_models
+export TAAR_ENSEMBLE_KEY=taar/ensemble/ensemble_weight.json
+export TAAR_WHITELIST_BUCKET=moz-fx-data-taar-pr-prod-e0f7-prod-models
+export TAAR_WHITELIST_KEY=addon_recommender/only_guids_top_200.json.bz2
+export TAAR_LOCALE_BUCKET=moz-fx-data-taar-pr-prod-e0f7-prod-models
+export TAAR_LOCALE_KEY=taar/locale/top10_dict.json.bz2
+export TAAR_SIMILARITY_BUCKET=moz-fx-data-taar-pr-prod-e0f7-prod-models
+export TAAR_SIMILARITY_DONOR_KEY=taar/similarity/donors.json.bz2
+export TAAR_SIMILARITY_LRCURVES_KEY=taar/similarity/lr_curves.json.bz2
+export TAARLITE_GUID_COINSTALL_BUCKET=moz-fx-data-taar-pr-prod-e0f7-prod-models
+export TAARLITE_GUID_COINSTALL_KEY=taar/lite/guid_coinstallation.json.bz2
+export TAARLITE_GUID_RANKING_KEY=taar/lite/guid_install_ranking.json.bz2
