+++ tests/test_collaborativerecommender.py
@@ -11,7 +11,9 @@ import numpy
 import fakeredis
 import mock
 import contextlib
-from taar.recommenders.redis_cache import TAARCache
+
+from taar.recommenders.cache import TAARCache
+from taar.recommenders.redis_cache import TAARCacheRedis
 
 
 from taar.recommenders.collaborative_recommender import CollaborativeRecommender
@@ -51,16 +53,16 @@ def mock_install_none_mock_data(ctx):
     we always get 404 errors.
     """
     with contextlib.ExitStack() as stack:
-        TAARCache._instance = None
+        TAARCacheRedis._instance = None
 
         stack.enter_context(
             mock.patch.object(
-                TAARCache, "_fetch_collaborative_item_matrix", return_value="",
+                TAARCacheRedis, "_fetch_collaborative_item_matrix", return_value="",
             )
         )
         stack.enter_context(
             mock.patch.object(
-                TAARCache, "_fetch_collaborative_mapping_data", return_value="",
+                TAARCacheRedis, "_fetch_collaborative_mapping_data", return_value="",
             )
         )
 
@@ -69,7 +71,7 @@ def mock_install_none_mock_data(ctx):
         # Patch fakeredis in
         stack.enter_context(
             mock.patch.object(
-                TAARCache,
+                TAARCacheRedis,
                 "init_redis_connections",
                 return_value={
                     0: fakeredis.FakeStrictRedis(db=0),
@@ -80,7 +82,9 @@ def mock_install_none_mock_data(ctx):
         )
 
         # Initialize redis
-        TAARCache.get_instance(ctx).safe_load_data()
+        cache = TAARCacheRedis.get_instance(ctx)
+        cache.safe_load_data()
+        ctx[TAARCache] = cache
         yield stack
 
 
@@ -109,17 +113,17 @@ def mock_install_mock_data(ctx):
         fake_mapping[str(java_hash)] = addon
 
     with contextlib.ExitStack() as stack:
-        TAARCache._instance = None
+        TAARCacheRedis._instance = None
         stack.enter_context(
             mock.patch.object(
-                TAARCache,
+                TAARCacheRedis,
                 "_fetch_collaborative_item_matrix",
                 return_value=fake_addon_matrix,
             )
         )
         stack.enter_context(
             mock.patch.object(
-                TAARCache,
+                TAARCacheRedis,
                 "_fetch_collaborative_mapping_data",
                 return_value=fake_mapping,
             )
@@ -130,7 +134,7 @@ def mock_install_mock_data(ctx):
         # Patch fakeredis in
         stack.enter_context(
             mock.patch.object(
-                TAARCache,
+                TAARCacheRedis,
                 "init_redis_connections",
                 return_value={
                     0: fakeredis.FakeStrictRedis(db=0),
@@ -141,7 +145,9 @@ def mock_install_mock_data(ctx):
         )
 
         # Initialize redis
-        TAARCache.get_instance(ctx).safe_load_data()
+        cache = TAARCacheRedis.get_instance(ctx)
+        cache.safe_load_data()
+        ctx[TAARCache] = cache
         yield stack
